Rename python namespace from keg to kiwi_keg
The project and namespace keg is already registered on the
python package index. Thus we need to rename the python
namespace as it cannot co-exist with an already registered
one
