+++ tox.ini
@@ -51,10 +51,10 @@ changedir=test/unit
 commands =
     bash -c 'cd ../../ && ./setup.py develop'
 
-    pytest --no-cov-on-fail --cov=keg --cov-report=term-missing
+    pytest --no-cov-on-fail --cov=kiwi_keg --cov-report=term-missing
 
 # disabled until we are finished
-#    pytest --no-cov-on-fail --cov=keg \
+#    pytest --no-cov-on-fail --cov=kiwi_keg \
 #        --cov-report=term-missing \
 #        --cov-fail-under=100 {posargs}
 
@@ -74,10 +74,10 @@ changedir=test/unit
 commands =
     bash -c 'cd ../../ && ./setup.py develop'
 
-    pytest --no-cov-on-fail --cov=keg --cov-report=term-missing
+    pytest --no-cov-on-fail --cov=kiwi_keg --cov-report=term-missing
 
 # disabled until we are finished
-#    pytest --doctest-modules --no-cov-on-fail --cov=keg \
+#    pytest --doctest-modules --no-cov-on-fail --cov=kiwi_keg \
 #        --cov-report=term-missing \
 #        --cov-fail-under=100 {posargs}
 
@@ -106,5 +106,5 @@ deps = {[testenv]deps}
 skip_install = True
 usedevelop = True
 commands =
-    flake8 --statistics -j auto --count {toxinidir}/keg
+    flake8 --statistics -j auto --count {toxinidir}/kiwi_keg
     flake8 --statistics -j auto --count {toxinidir}/test/unit
